Code review change - continue refactoring filters, separate classes to files
